+++ pkg/print/config.go
@@ -6,6 +6,7 @@ import (
 
 	configtypes "github.com/replicatedhq/kots/pkg/kotsadmconfig/types"
 	"github.com/replicatedhq/kots/pkg/logger"
+	"github.com/replicatedhq/troubleshoot/pkg/preflight"
 )
 
 func ConfigValidationErrors(log *logger.CLILogger, groupValidationErrors []configtypes.ConfigGroupValidationError) {
@@ -25,4 +26,28 @@ func ConfigValidationErrors(log *logger.CLILogger, groupValidationErrors []confi
 
 	log.FinishSpinnerWithError()
 	log.Errorf(sb.String())
+}
+
+func PreflightErrors(log *logger.CLILogger, results []*preflight.UploadPreflightResult) {
+	var s strings.Builder
+	s.WriteString("\nPreflight check results (state - title - message)")
+	for _, result := range results {
+		s.WriteString(fmt.Sprintf("\n%s - %q - %q", preflightState(result), result.Title, result.Message))
+	}
+	log.Info(s.String())
+	// log.Errorf(sb.String())
+}
+
+func preflightState(r *preflight.UploadPreflightResult) string {
+	if r.IsFail {
+		return "FAIL"
+	}
+	if r.IsWarn {
+		return "WARN"
+	}
+	if r.IsPass {
+		return "PASS"
+	}
+	// We should never get here
+	return "UNKNOWN"
 }
